Generate new SSH key for admin account
The Terraform provider used to generate the key pair stores the private
key unencrypted in the terraform state file
(https://registry.terraform.io/providers/hashicorp/tls/latest/docs/resources/private_key).

The provider is indented mainly for disposable development environments.
The use here should be fine as long as the state is not shared (for
example in terraform cloud). A note to this effect has been added to the
README.